Fixing bugs caused by trying to pack stats into shorts rather than ints (wow, pages can recieve a lot of links - too many to count with shorts)

+++ src/org/wikipedia/miner/db/DbAnchor.java
@@ -77,8 +77,8 @@ public class DbAnchor {
 			Arrays.sort(mergedSenses) ;
 		}
 			
-		totalLinks = (short)(totalLinks + anchor.totalLinks) ;
-		distinctLinks = (short)(distinctLinks + anchor.distinctLinks) ;
+		totalLinks = totalLinks + anchor.totalLinks ;
+		distinctLinks = distinctLinks + anchor.distinctLinks ;
 		totalReferences = totalReferences + anchor.totalReferences ;
 		distinctReferences = distinctReferences + anchor.distinctReferences ;
 		senses = mergedSenses ;
